Remove dtm argument to corpusCa()
When omitted, since it was set to NULL, the dtm was computed from scratch.
This means all changes done to the global dtm object were never used by
the GUI.

Always use the global dtm object, the argument wasn't used anywhere.
Only compute the dtm if it doesn't exist, for mere convenience.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@4 edb9625f-4e0d-4859-8d74-9fd3b1da38cb