Add corpus dissimilarity dialog
Fix a small glitch in showCorpusClustering: the column weights were
computed using the sum of the dtm _without_ the terms that do not
appear in the considered cluster. We do not need these terms in other
steps of the computation, but we need them here for the total sum.
This ensures we get exactly the same result as for dissimilarity, when
crossing documents clusters with sparsity = 1.

Also the formula to compute distance the other way round: instead of
multiplying by inverse of column probabilities, divide by these probs.
This does not change anything, but feels more natural.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@184 edb9625f-4e0d-4859-8d74-9fd3b1da38cb