Improve time series feature
Handle irregular time series better: we need to aggregate the original
data in case it is not even weakly regular (i.e. the smallest difference
between times is not a divisor of all the differences between times).
Also detect more correctly what is the appropriate minimum time unit,
use it, and report it on the plot.

Also allow choosing between number of documents and % of documents. Fix
detection of series length < window length. Do not close dialog when
plotting. Update DESCRIPTION.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@191 edb9625f-4e0d-4859-8d74-9fd3b1da38cb