Factiva: Work around encoding issues in Windows for HTML files
This does not fix problems with XML format.

xmlParse() and htmlParse() do not set Encoding() correctly on the
resulting strings. This is not visible when UTF-8 is the default
encoding, but on Windows it's not the case, and accentuated
characters do not appear correctly.

We cannot fix this in XML from tm.plugin.factiva, and since the best
solution would be to fix this in XML, let it that way for now. As a
side note, ReutersSource() from tm is also affected.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@206 edb9625f-4e0d-4859-8d74-9fd3b1da38cb