Add Twitter support
This required fixing a few things in the management of time series:
since tweets come with a timestamp in hours-minutes-seconds, and
seconds are not very useful (they make table() eat much RAM since
every timestamp is unique in seconds), we better ignore them by
default. This means the strptime() format should ignore them,
which in turn implies we have to support cases where the original
timestamps are unique, but the POSIXct objects are not.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@246 edb9625f-4e0d-4859-8d74-9fd3b1da38cb