Add "All files" pattern to file selection dialogs and improve order
It's always good to be able to select any file, in case the extension
is missing, and because in some cases on Windows the last pattern is selected
by default instead of the first. For spreadsheet-like import, we assume
the file is in TSV format, so it's not very useful...

Put the default pattern at the start of the list, as it is the order
that is supposed to work everywhere.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@276 edb9625f-4e0d-4859-8d74-9fd3b1da38cb