Avoid opening Commander window twice on load
After discussing this with John Fox, a solution was found, which
required a few changes in Rcmdr itself so that we can run it
without adding it to Depends, only to Imports. Because Rcmdr
redefines some functions from tcltk, a warning would appear on
install: to remove it, only import non-conflicting symbols we use
from tcltk. This requires Rcmdr >= 1.9-3 (not yet on CRAN).

Also fix indentation in zzz.R, and rename it to startup.R.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@369 edb9625f-4e0d-4859-8d74-9fd3b1da38cb