Fix dictionary after subsetting corpus
Subsetting dtm discards its attributes, which means we did not
have the language information: this created a Java error when
calling Snowball with an empty code.

More generally, the list of words with their frequencies is extracted
on import, and we need to recompute it after removing documents: factor
out text processing to a new processTexts() function, and call it when
subsetting texts only to extract words.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@380 edb9625f-4e0d-4859-8d74-9fd3b1da38cb