Improve HTML output of CA scree plot
Better output a data frame with numeric columns except for the stars
than a character matrix filled manually. This allows for automatic
rounding and using the right decimal separator. Do not show all
dimensions, but only up to 15, so that the last eigenvalues are not
too small: this means we do not show too much noise, and
keeping only 2 significant digits now has an effect.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@404 edb9625f-4e0d-4859-8d74-9fd3b1da38cb