Improve setting busy cursor on import and when plotting CA
On import, the cursor had to be set after closing the variables
selection dialog. Also show it when plotting CA as this can be
slow; the busy cursor should also be set on the plot dialog and
on the CA dialog, but this would better be fixed in Rcmdr now.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@471 edb9625f-4e0d-4859-8d74-9fd3b1da38cb