Improve display of most frequent terms
Sort them by t-value, which is more precise than the p-value: the latter is
derived from the former, and can sometimes be equal to 0 for different t-values.
Also change the sort order to show most negative values at the end, which is more
natural. To make this more obvious, add a separation row.

git-svn-id: svn+ssh://scm.r-forge.r-project.org/svnroot/r-temis@589 edb9625f-4e0d-4859-8d74-9fd3b1da38cb