Improve splitTexts() and use VCorpus to fix tm 0.7 failures
Documents from DirSource are no longer a vector of paragraphs, so we
need to split texts on newline characters. The new version is more robust
since it also works with the old structure.
Also improve computing names of document chunks, starting at 1.

SimpleCorpus is faster than VCorpus, but it does not support setting
document IDs, which we need to do to create chunks and to allow the user to load names
from a data frame in setCorpusVariables(). It also suffers from a bug
which loses encoding when computing DocumentTermMatrix.