Fix terms dictionary when importing it from CSV
The order of columns was different from the standard case, and later code
relied on it. Make it more robust by using names instead, and use the
same order everywhere.

Also avoid writing rownmames to CSV since these are redundant with the
first column, and use stringsAsFactors=FALSE on import.