Fix Tcl error when showing CA results multiple times
The error happened after showing results once, selecting an entry in the list,
and then showing the results again with a lower number of shown items.
This was because there was no selection in the list box, and emptying it
triggered the callback which then tried to find a mark in the text view
with no number.