Handle non-numeric 'page' metadata entries
Some newspapers (e.g. the New York Times and the Wall Street Journal)
number their pages alphanumerically by section and by number (e.g. A6).
The existing code attempted to strip out the alphabetic parts, which
are meaningful in this context, and also caused an error if there was
no numeric part to extract.

Replace this with code which simply checks for NA and otherwise saves
the Factiva PG entry into the metadata list as a string.

Note that this changes the typeof(meta(x, "page")) from double to
character, which may not be compatible with all downstream uses.