Ensure FactivaSource IDs are unique
`readFactivaHTML` extracts useful information from the AN field to
construct an id. Unfortunately, this isn't guaranteed to be unique
(unlike the AN field itself) especially if a large number of
articles from the same day are ingested. Non-unique ids are
causing problems with corpus creation: only one article per ID is
added to the corpus, but duplicated.

This patch amends the original code to simply use the AN field
directly where possible (with the original fallback of a random
10 letter id where necessary). This is a little less elegant but
should be guaranteed to work.